kindle-to-PDF/requirements.txt
pyautogui==0.9.54
Pillow>=10.2.0
reportlab>=4.0.0 